rename NightWatch -> xdrip/dexdrip in module wear
