+++ wear/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.dexdrip.stephenblack.nightwatch" >
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.eveningoutpost.dexdrip" >
 
     <uses-feature android:name="android.hardware.type.watch" />
     <uses-permission android:name="com.google.android.permission.PROVIDE_BACKGROUND" />
@@ -16,7 +16,7 @@
         <service
             android:name=".Home"
             android:allowEmbedded="true"
-            android:label="NightWatch"
+            android:label="xDrip"
             android:permission="android.permission.BIND_WALLPAPER">
             <meta-data android:name="android.service.wallpaper"
                        android:resource="@xml/watch_face"/>
@@ -30,7 +30,7 @@
         <service
             android:name=".LargeHome"
             android:allowEmbedded="true"
-            android:label="NightWatch(Large)"
+            android:label="xDrip(Large)"
             android:permission="android.permission.BIND_WALLPAPER">
             <meta-data android:name="android.service.wallpaper"
                        android:resource="@xml/watch_face"/>
@@ -44,7 +44,7 @@
         <service
             android:name=".BIGChart"
             android:allowEmbedded="true"
-            android:label="NightWatch(BigChart)"
+            android:label="xDrip(BigChart)"
             android:permission="android.permission.BIND_WALLPAPER">
             <meta-data android:name="android.service.wallpaper"
                 android:resource="@xml/watch_face"/>
@@ -58,7 +58,7 @@
         <service
             android:name=".CircleWatchface"
             android:allowEmbedded="true"
-            android:label="NightWatch(Circle)"
+            android:label="xDrip(Circle)"
             android:permission="android.permission.BIND_WALLPAPER">
             <meta-data android:name="android.service.wallpaper"
                 android:resource="@xml/watch_face"/>
