+++ wear/src/main/java/com/dexdrip/stephenblack/dexdrip/BgWatchData.java
@@ -1,4 +1,4 @@
-package com.dexdrip.stephenblack.nightwatch;
+package com.dexdrip.stephenblack.dexdrip;
 
 /**
  * Created by stephenblack on 1/7/15.
