+++ wear/src/main/java/com/dexdrip/stephenblack/dexdrip/Home.java
@@ -1,4 +1,4 @@
-package com.dexdrip.stephenblack.nightwatch;
+package com.dexdrip.stephenblack.dexdrip;
 
 import android.content.Context;
 import android.graphics.Color;
@@ -6,19 +6,19 @@ import android.view.LayoutInflater;
 
 import com.ustwo.clockwise.WatchMode;
 
-public class LargeHome extends BaseWatchFace {
+import lecho.lib.hellocharts.util.Utils;
+
+public class Home extends BaseWatchFace {
 
     @Override
     public void onCreate() {
         super.onCreate();
         LayoutInflater inflater = (LayoutInflater) getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-        layoutView = inflater.inflate(R.layout.activity_home_large, null);
+        layoutView = inflater.inflate(R.layout.activity_home, null);
         performViewSetup();
     }
 
-
-    @Override
-    protected void setColorDark(){
+    protected void setColorDark() {
         mTime.setTextColor(Color.WHITE);
         mRelativeLayout.setBackgroundColor(Color.BLACK);
         mLinearLayout.setBackgroundColor(Color.WHITE);
@@ -46,21 +46,28 @@ public class LargeHome extends BaseWatchFace {
         } else {
             mUploaderBattery.setTextColor(Color.RED);
         }
-
         mRaw.setTextColor(Color.BLACK);
+        if (chart != null) {
+            highColor = Color.YELLOW;
+           lowColor = Color.RED;
+            midColor = Color.WHITE;
+            singleLine = false;
+            pointSize = 2;
+            setupCharts();
+        }
+
     }
 
 
-
-    @Override
     protected void setColorBright() {
+
         if (getCurrentWatchMode() == WatchMode.INTERACTIVE) {
             mRelativeLayout.setBackgroundColor(Color.WHITE);
             mLinearLayout.setBackgroundColor(Color.BLACK);
             if (sgvLevel == 1) {
-                mSgv.setTextColor(Color.YELLOW);
-                mDirection.setTextColor(Color.YELLOW);
-                mDelta.setTextColor(Color.YELLOW);
+                mSgv.setTextColor(Utils.COLOR_ORANGE);
+                mDirection.setTextColor(Utils.COLOR_ORANGE);
+                mDelta.setTextColor(Utils.COLOR_ORANGE);
             } else if (sgvLevel == 0) {
                 mSgv.setTextColor(Color.BLACK);
                 mDirection.setTextColor(Color.BLACK);
@@ -84,9 +91,17 @@ public class LargeHome extends BaseWatchFace {
             }
             mRaw.setTextColor(Color.WHITE);
             mTime.setTextColor(Color.BLACK);
+            if (chart != null) {
+                highColor = Utils.COLOR_ORANGE;
+                midColor = Color.BLUE;
+                lowColor = Color.RED;
+                singleLine = false;
+                pointSize = 2;
+                setupCharts();
+            }
         } else {
             mRelativeLayout.setBackgroundColor(Color.BLACK);
-            mLinearLayout.setBackgroundColor(Color.LTGRAY);
+            mLinearLayout.setBackgroundColor(Color.WHITE);
             if (sgvLevel == 1) {
                 mSgv.setTextColor(Color.YELLOW);
                 mDirection.setTextColor(Color.YELLOW);
@@ -100,11 +115,20 @@ public class LargeHome extends BaseWatchFace {
                 mDirection.setTextColor(Color.RED);
                 mDelta.setTextColor(Color.RED);
             }
-
+            mRaw.setTextColor(Color.BLACK);
             mUploaderBattery.setTextColor(Color.BLACK);
             mTimestamp.setTextColor(Color.BLACK);
-            mRaw.setTextColor(Color.BLACK);
+
             mTime.setTextColor(Color.WHITE);
+            if (chart != null) {
+                highColor = Color.YELLOW;
+                midColor = Color.WHITE;
+                lowColor = Color.RED;
+                singleLine = true;
+                pointSize = 2;
+                setupCharts();
+            }
         }
+
     }
-}
+    }
