+++ wear/src/main/java/com/dexdrip/stephenblack/dexdrip/NWPreferences.java
@@ -1,4 +1,4 @@
-package com.dexdrip.stephenblack.nightwatch;
+package com.dexdrip.stephenblack.dexdrip;
 
 
 import android.os.Bundle;
